WORKING WITH HARDWARE!!!!! fully debugged communication with python script (adapted python script, found 1-off error, found peripheral mapping error, found error with assigning ISR's after InitUART, ...) added timeout in case communication is lost debugged issue with configuration bit and PPS module (thank you Taylor) debugged several macros that are active low (not active high)

+++ Avatar_Micro_Robot/Robot_V2_Accessories/Robotic_Arm_V2/Base/trunk/src/PWM.c
@@ -54,7 +54,7 @@ void InitPWM(unsigned char pin, unsigned int us) {
 }
 
 void UpdateDutyCycle(unsigned char duty_cycle) {
-	OC2R = (unsigned int) ((float)duty_cycle / 100 * (float)period);
+	OC2R = (((float)duty_cycle / 100.0) * (float)period);
 }
